% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossings.R
\name{Crossings}
\alias{Crossings}
\title{Crossings Interaction of Factors}
\usage{
Crossings(...)
}
\arguments{
\item{\dots}{Two factors}
}
\value{
For two factors of \code{n} levels, returns a binary indicator
matrix of \code{n*n} rows and \code{n-1} columns.
}
\description{
Given two ordered factors in a square, n x n frequency table,
\code{Crossings} creates an n-1 column matrix corresponding to different
degrees of difficulty in crossing from one level to the next, as described
by Goodman (1972).
}
\details{
Instead of treating all mobility as equal, this model posits that the difficulty
of moving between categories increases with the number of boundaries (or "crossings")
that must be crossed, and that associations between categories decrease
with their separation.
}
\examples{

data(Hauser79)
# display table
structable(~Father + Son, data=Hauser79)

hauser.indep <- gnm(Freq ~ Father + Son,
                    data=Hauser79,
                    family=poisson)

hauser.CR <- update(hauser.indep,
                    ~ . + Crossings(Father,Son))
LRstats(hauser.CR)

hauser.CRdiag <- update(hauser.indep,
                        ~ . + Crossings(Father,Son) + Diag(Father,Son))
LRstats(hauser.CRdiag)

# what does Crossings do?
cr <- with(Hauser79, Crossings(Father, Son))
head(cr)
# Show the codings for varying Crossings levels
matrix(cr[,1], nrow=5)
matrix(cr[,2], nrow=5)
matrix(cr[,3], nrow=5)
matrix(cr[,4], nrow=5)

}
\references{
Goodman, L. (1972).  Some multiplicative models for the analysis
of cross-classified data. In: \emph{Proceedings of the Sixth Berkeley
Symposium on Mathematical Statistics and Probability}, Berkeley, CA:
University of California Press, pp. 649-696.
}
\seealso{
\code{\link[stats]{glm}}, \code{\link[gnm]{gnm}} for model fitting
functions for frequency tables;
\code{\link[gnm]{Diag}}, \code{\link[gnm]{Mult}}, \code{\link[gnm]{Symm}},
\code{\link[gnm]{Topo}} for similar extensions to terms in model formulas.
}
\author{
Michael Friendly and Heather Turner
}
\keyword{manip}
\keyword{models}
