% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfac.R
\name{cutfac}
\alias{cutfac}
\title{Cut a Numeric Variable to a Factor}
\usage{
cutfac(x, breaks = NULL, q = 10)
}
\arguments{
\item{x}{a numeric vector which is to be converted to a factor by cutting}

\item{breaks}{either a numeric vector of two or more unique cut points or a
single number (greater than or equal to 2) giving the number of intervals
into which \code{x} is to be cut.}

\item{q}{the number of quantile groups used to define \code{breaks}, if that
has not been specified.}
}
\value{
A \code{\link[base]{factor}} corresponding to \code{x} is returned
}
\description{
\code{cutfac} acts like \code{\link[base]{cut}}, dividing the range of
\code{x} into intervals and coding the values in \code{x} according in which
interval they fall. However, it gives nicer labels for the factor levels and
by default chooses convenient breaks among the values based on deciles.
}
\details{
It is particularly useful for plots in which one wants to make a numeric
variable discrete for the purpose of getting boxplots, spinograms or mosaic
plots.

By default, \code{\link[base]{cut}} chooses breaks by equal lengths of the
range of \code{x}, whereas \code{cutfac} uses \code{\link[stats]{quantile}}
to choose breaks of roughly equal count.
}
\references{
Friendly, M. and Meyer, D. (2016).  \emph{Discrete Data Analysis
with R: Visualization and Modeling Techniques for Categorical and Count
Data}.  Boca Raton, FL: Chapman & Hall/CRC. \url{http://ddar.datavis.ca}.
}
\seealso{
\code{\link[base]{cut}}, \code{\link[stats]{quantile}}
}
\author{
Achim Zeileis
}
\keyword{manip}
