% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorkerSat}
\alias{WorkerSat}
\title{Worker Satisfaction Data}
\format{
A frequency data frame with 8 observations on the following 4
variables, representing the 2 x 2 x 2 classification of 715 cases.
\describe{ \item{list("Manage")}{Quality of management, an ordered factor
with levels \code{bad} < \code{good}} \item{list("Super")}{Supervisor
satisfaction, an ordered factor with levels \code{low} < \code{high}}
\item{list("Worker")}{Worker job satisfaction, an ordered factor with levels
\code{low} < \code{high}} \item{list("Freq")}{a numeric vector} }
}
\source{
Originally from \url{https://online.stat.psu.edu/stat504/lesson/10/}
}
\description{
Blue collar workers job satisfaction from large scale investigation in
Denmark in 1968 (Andersen, 1991).
}
\examples{

data(WorkerSat)

worker.tab <- xtabs(Freq ~ Worker + Super + Manage, data=WorkerSat)
fourfold(worker.tab)
mosaic(worker.tab, shade=TRUE)


}
\references{
Andersen, E. B. (1991) Statistical Analysis of Categorical Data,
2nd Ed., Springer-Verlag.
}
\keyword{datasets}
