% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglin-utilities.R
\name{joint}
\alias{joint}
\alias{conditional}
\alias{mutual}
\alias{saturated}
\alias{markov}
\alias{loglin2formula}
\alias{loglin2string}
\title{Loglinear Model Utilities
These functions generate lists of terms to specify a loglinear model
in a form compatible with loglin and provide for conversion to an
equivalent loglm specification.  They allow for a more conceptual
way to specify such models.
models of joint independence, of some factors wrt one or more other factors}
\source{
Code from Henrique Dallazuanna, <wwwhsd@gmail.com>, R-help 7-4-2013
}
\usage{
joint(nf, table = NULL, factors = 1:nf, with = nf)

conditional(nf, table = NULL, factors = 1:nf, with = nf)

mutual(nf, table = NULL, factors = 1:nf)

saturated(nf, table = NULL, factors = 1:nf)

markov(nf, factors = 1:nf, order = 1)

loglin2formula(x, env = parent.frame())

loglin2string(x, brackets = c("[", "]"), sep = ",", collapse = " ", abbrev)
}
\arguments{
\item{nf}{number of factors for which to generate model}

\item{table}{a contingency table used for factor names, typically the output from \code{\link[base]{table}}}

\item{factors}{names of factors used in the model when \code{table} is not specified}

\item{with}{indices of the factors against which others are considered conditionally independent}

\item{order}{order of the markov chain}

\item{x}{a list of terms in a loglinear model, such as returned by \code{joint}, \code{conditional}, \dots}

\item{env}{environment in which to evaluate the formula}

\item{brackets}{characters to use to surround model terms.  Either a single character string containing two characters
or a character vector of length two.}

\item{sep}{characters used to separate factor names within a term}

\item{collapse}{characters used to separate terms}
}
\description{
Loglinear Model Utilities
These functions generate lists of terms to specify a loglinear model
in a form compatible with loglin and provide for conversion to an
equivalent loglm specification.  They allow for a more conceptual
way to specify such models.
models of joint independence, of some factors wrt one or more other factors

models of conditional independence of some factors wrt one or more other factors

models of mutual independence of all factors

saturated model: highest-order interaction

markov models of a given order

convert a loglin model to a model formula for loglm

convert a loglin model to a string, using bracket notation for the high-order terms
}
