\name{vcdExtra-deprecated}
\alias{vcdExtra-deprecated}
%\alias{summarise.glm}
%\alias{summarise.glm}
%\alias{summarise.glmlist}
%\alias{summarise.loglm}
%\alias{summarise.loglmlist}
%
\title{Deprecated Functions in vcdExtra Package}
%
\description{
  These functions are provided for compatibility with older versions of the \pkg{vcdExtra} package only, 
  and may be removed eventually.
}
%
%\usage{
%summarise(...)
%summarise.glm(...)
%summarise.glmlist(...)
%summarise.loglm(...)
%summarise.loglmlist(...)
%}
%
%\arguments{
%  \item{\dots}{pass arguments down.}
%}
%
%\details{
%\code{summarise.*} have been replaced by \code{\link{LRstats}} functions.
%}
